+++ tests/testthat/test-joins.R
@@ -2,33 +2,20 @@ context("joins")
 
 
 test_that("joining data tables returns data tables (#470) and does not modify them (#659)", {
-  a <- data.table(x = c(1, 1, 2, 3), y = 4:1)
-  b <- data.table(x = c(1, 2, 2, 4), z = 1:4)
+  x <- data.table(x = c(1, 1, 2, 3), y = 4:1)
+  y <- data.table(x = c(1, 2, 2, 4), z = 1:4)
 
-  test_join <- function(join_fun, ak, bk) {
-    data.table::setkeyv(a, ak)
-    data.table::setkeyv(b, bk)
-    ac <- data.table::copy(a)
-    bc <- data.table::copy(b)
+  join_funs <- list(left_join, semi_join, anti_join)
+  for (join_fun in join_funs) {
+    xprime <- data.table::copy(x)
+    yprime <- data.table::copy(y)
 
-    out <- join_fun(a, b, "x")
-    expect_is(out, "data.table")
-    expect_equal(a, ac)
-    expect_equal(b, bc)
-  }
-
-  for (ak in names(a)) {
-    for (bk in names(b)) {
-      test_join(left_join, ak, bk)
-      test_join(semi_join, ak, bk)
-      test_join(right_join, ak, bk)
-      test_join(full_join, ak, bk)
-      test_join(inner_join, ak, bk)
-      test_join(anti_join, ak, bk)
-    }
+    out <- join_fun(xprime, yprime, "x")
+    expect_s3_class(out, "data.table")
+    expect_equal(xprime, x)
+    expect_equal(yprime, y)
   }
 })
-
 
 test_that("joining data tables returns same result as dplyr", {
   a_dt <- data.table(x = c(1, 1, 2, 3), y = 4:1)